dataman: split out file, ram, and ram_flash operations
